// functions/src/models/institution.ts
import { Contact } from "./contact";

export class Institution {
    id: string;
    place: string;
    subordinationLevel: string;
    placeType: string;
    fullName: string;
    fullAddress: string;
    institutionType: string;
    email: string;
    phone: string;
    fax: string;
    site: string;
    legalStatus: boolean;
    stateRegisterCode: string;
    classifierObjectCode: string;
    ownership: string;
    legalFormCode: string;
    headDoctor: Contact;
    headDoctorSecretary: Contact;
    receptionOffice: boolean;
    receptionOfficePhone: string;
    registryOffice: boolean;
    registryOfficePhone: string;
    population: number;
    totalPopulation: number;
    medicalСare: string;
    іnstitutionType: string;
    medicalAidTypes: string[];
    declaredAssistanceForms: string;
    license: boolean;
    licenseNumber: string;
    licenseDate: Date;
    startLicenseValidity: Date;
    endLicenseValidity: Date;
    accreditationCategoryType: string;
    accreditationСategoryNumber: string;
    lastAccreditation: Date;
    startAccreditationValidity: Date;
    endAccreditationValidity: Date;
    hospitalBedsNumber: number;
    hospitalCapacity: number;
    regularDoctorNumber: number;
    busyDoctorNumber: number;
    individualsDoctorNumber: number;
    middleRegularPersonalNumber: number;
    middleBusyPersonalNumber: number;
    middleIndividualsPersonalNumber: number;
    otherRegularPersonalNumber: number;
    otherBusyPersonalNumber: number;
    otherIndividualsPersonalNumber: number;
    totalRegularPersonalNumber: number;
    totalBusyPersonalNumber: number;
    totalIndividualsPersonalNumber: number; 
    heatingType: string;
    typicalBuilding: boolean;
    coldWaterSupply: boolean;
    hotWaterSupply: boolean; 
    internetSupply: boolean;
    equipment: number;
    medicamentEquipment: number;
    vehiclesNeed: number;
    vehiclesReality: number;
    vehiclesWay: number;
    computerEquipmentNeed: number;
    computerEquipmentReality: number;
    controlledInstitutions: Institution[];
    controlledBy: string;
    type: number;
}